requirements.txt
flask
scikit-learn
openai
flask-cors
gunicorn
python-dotenv